heap and vector nao testados

+++ heap.h
@@ -7,6 +7,6 @@ typedef struct _heap Heap;
 
 Heap* criaHeap();
 
-void insereHeap (Heap*, int idVertice, int distancia); //-1 se não tiver ligamento com o vértice principal
+void insereHeap (Heap*, int idVertice, float distancia); //-1 se não tiver ligamento com o vértice principal
 
 #endif
